+++ source/InjectHooksMain.cpp
@@ -397,8 +397,7 @@
 
 #include "platform/win/VideoPlayer/VideoPlayer.h"
 #include "platform/win/VideoMode.h"
-#include "platform/win/win.h"
-#include "platform/platform.h"
+#include "platform/win/Platform.h"
 
 #include "app/app.h"
 #include <RealTimeShadowManager.h>
