+++ source/app/platform/win/Platform.h
