+++ source/app/platform/win/WinMain.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void InjectWinMainStuff();
